Preserve the authored precision of USD geometry sizes and poses.
The size of a UsdGeom.Cube and the composed local-to-world transform of every prim are authored in double
precision by USD, but were both narrowed to single precision on the way in, costing every box extent and
every geom pose their last digits for no reason. They are now kept as authored and narrowed only once the
value reaches the entity, at the simulation precision.

UsdPhysics declares every joint attribute as single precision, so an anchor, a limit or a drive gain
authored in double genuinely cannot be recovered from a USD file. The MJCF-vs-USD joint comparison is
therefore held to USD_PHYSICS_TOL rather than the session tolerance. That comparison was also vacuous for
the joint poses, which the fixture authored as zeros: the prismatic and revolute joints now carry real
anchors, expressed in both body frames on the USD side.

+++ genesis/utils/usd/usd_context.py
@@ -248,7 +248,9 @@ class UsdContext:
         Compute the local-to-world transformation matrix for a prim.
         """
         transform = self._xform_cache.GetLocalToWorldTransform(prim)
-        T_usd = np.asarray(transform, dtype=np.float32)  # translation on the bottom row
+        # USD composes transforms in double precision, so the matrix is kept as authored and only narrowed to the
+        # simulation precision once the pose reaches the entity: rounding it here costs the pose its last digits.
+        T_usd = np.asarray(transform)  # translation on the bottom row
         if self._is_yup:
             T_usd @= mu.Y_UP_TRANSFORM
         T_usd[:, :3] *= self._meter_scale
